Make `gix::Repository` extension available in `oxidize`.
It has minimal dependencies and can be used everywhere.
