+++ crates/gitbutler-branch-actions/src/branch.rs
@@ -7,11 +7,11 @@ use gitbutler_branch::BranchIdentity;
 use gitbutler_branch::ReferenceExtGix;
 use gitbutler_command_context::CommandContext;
 use gitbutler_diff::DiffByPathMap;
-use gitbutler_oxidize::{git2_to_gix_object_id, gix_to_git2_oid};
+use gitbutler_oxidize::{git2_to_gix_object_id, gix_to_git2_oid, GixRepositoryExt};
 use gitbutler_project::access::WorktreeReadPermission;
 use gitbutler_reference::normalize_branch_name;
 use gitbutler_reference::RemoteRefname;
-use gitbutler_repo::{GixRepositoryExt, RepositoryExt as _};
+use gitbutler_repo::RepositoryExt as _;
 use gitbutler_serde::BStringForFrontend;
 use gitbutler_stack::{Stack as GitButlerBranch, StackId, Target};
 use gix::object::tree::diff::Action;
