+++ crates/gitbutler-branch-actions/src/branch_manager/branch_removal.rs
@@ -5,11 +5,10 @@ use git2::Commit;
 use gitbutler_branch::BranchExt;
 use gitbutler_commit::commit_headers::CommitHeadersV2;
 use gitbutler_oplog::SnapshotExt;
-use gitbutler_oxidize::git2_to_gix_object_id;
 use gitbutler_oxidize::gix_to_git2_oid;
+use gitbutler_oxidize::{git2_to_gix_object_id, GixRepositoryExt};
 use gitbutler_project::access::WorktreeWritePermission;
 use gitbutler_reference::{normalize_branch_name, ReferenceName, Refname};
-use gitbutler_repo::GixRepositoryExt;
 use gitbutler_repo::RepositoryExt;
 use gitbutler_repo::SignaturePurpose;
 use gitbutler_repo_actions::RepoActionsExt;
